requirements.txt
hydra-core>=1.1
termcolor
omegaconf
gym
wandb
modal
trimesh
scipy
matplotlib
imageio
imageio-ffmpeg
